% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{accessors}
\alias{accessors}
\alias{$.multiverse}
\alias{multiverse_table}
\alias{multiverse_table.default}
\alias{multiverse_table.multiverse}
\alias{code}
\alias{code.default}
\alias{code.multiverse}
\alias{parameters}
\alias{parameters.default}
\alias{parameters.multiverse}
\alias{conditions}
\alias{conditions.default}
\alias{conditions.multiverse}
\alias{from_universe_i}
\title{Accessing contents of the multiverse object}
\usage{
\method{$}{multiverse}(multiverse, name)

multiverse_table(multiverse)

\method{multiverse_table}{default}(multiverse)

\method{multiverse_table}{multiverse}(multiverse)

code(multiverse)

\method{code}{default}(multiverse)

\method{code}{multiverse}(multiverse)

parameters(multiverse)

\method{parameters}{default}(multiverse)

\method{parameters}{multiverse}(multiverse)

conditions(multiverse)

\method{conditions}{default}(multiverse)

\method{conditions}{multiverse}(multiverse)

from_universe_i(multiverse, idx, name)
}
\arguments{
\item{multiverse}{Object of class multiverse}

\item{name}{a variable name}

\item{idx}{index of the universe in the multiverse (corresponds to the row in the table)}
}
\description{
A multiverse object contains several \strong{Object variables}. These can be accessed using convenient functions.
Variables from the analysis that is being performed within the multiverse can be accessed using the `$`.
Object variables such as the `code`, the `multiverse table`, the `parameters` and the `conditions` can be accessed using respective functions
}
