% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blazr-package.R
\docType{package}
\name{blazr-package}
\alias{blazr}
\alias{blazr-package}
\title{blazr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-staceans/blazr}
  \item \url{https://r-staceans.github.io/blazr/}
  \item Report bugs at \url{https://github.com/r-staceans/blazr/issues}
}

}
\author{
\strong{Maintainer}: Jackson Hoffart \email{jackson.hoffart@gmail.com} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID})

}
\keyword{internal}
