% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_default}
\alias{execute_default}
\title{Execute parts of, or the entire multiverse}
\usage{
execute_default(multiverse, N = NA)
}
\arguments{
\item{multiverse}{The multiverse object}

\item{N}{Override the default analysis and instead perform the N-th analysis from the multiverse table}

\item{.vec}{A vector specifying the range of analysis paths from the multiverse to be executed. Defaults to \code{\link[base]{NA}}
which indicates the complete multiverse is to be executed.}
}
\description{
These are functions which allow the user to execute the multiverse. The user can choose to either execute the default analysis
using the execute_default() or parts or whole of the multiverse using the execute_multiverse(). These functions allow the user 
to interactively inspect the code.
}
