% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_json.R
\name{export_2_json}
\alias{export_2_json}
\alias{export_dist_2_json}
\title{Exporting results from a multiverse analysis to JSON}
\usage{
export_2_json(x, term, mean, sd, dist, filename)

export_dist_2_json(x, term, dist, filename)
}
\arguments{
\item{x}{a tidy tibble or data frame which contains summary statistics or distributional information 
of each regression coefficient parameter}

\item{term}{column in the data frame, x, which contains the name of the coefficients}

\item{mean}{column in the data frame, x, which contains the mean estimate for each coefficient}

\item{sd}{column in the data frame, x, which contains the standard error estimate for each coefficient}

\item{dist}{column in the data frame, x, which contains vectorised distributions---an object of class 
`distribution` for each coefficient}

\item{filename}{filename on disk (as a character string)}
}
\value{
a data frame or a JSON file
}
\description{
Exports the results of the multiverse analysis to JSON in a format which is compatible with the multiverse visualisation tool
}
\examples{
\donttest{
library(dplyr)
library(tidyr)

M = multiverse()

inside(M, {
  df = tibble(
    x = rnorm(100),
    y = x * 0.5 + rnorm(100, 0, 2)
 )
 
 # alternatives to remove outlier
 df.filtered = df \%>\%
   filter(
     branch(outlier_exclusion,
        "2SD" ~ abs(y - mean(y)) > 2*sd(y),
        "3SD" ~ abs(y - mean(y)) > 3*sd(y)
     )
 )
  
  fit = lm(y ~ x, data = df)
  res = broom::tidy(fit)
})

execute_multiverse(M)

multiverse::expand(M) \%>\%
  extract_variables(res) \%>\%
  unnest(res) \%>\%
  export_2_json(term, estimate, std.error) 
} 


}
