% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-code.R
\name{get_universe}
\alias{get_universe}
\title{Code corresponding to a single analysis}
\usage{
get_universe(M, .assgn = list())
}
\arguments{
\item{M}{The multiverse object with some code passed to it}

\item{.assgn}{A list containing the assignments for each defined parameter in the multiverse}
}
\description{
For a particular parameter assignment (i.e. one set of values that each defined parameter takes), 
this function rewrites the code passed into the multiverse to output the corresponding code for that
set of parameter values --- the analysis for a single universe. This function is primarily going to 
be called by other functions, and perhaps not going to be as useful to the user for anything other than
inspecting the rewritten code.
}
