% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse.R
\name{is_multiverse}
\alias{is_multiverse}
\title{Test if the object is a multiverse}
\usage{
is_multiverse(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the `multiverse` class.
}
\description{
This function returns `TRUE` for objects of class multiverse,
and `FALSE` for all other objects.
}
