% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_as_emar.R
\name{knit_as_emar}
\alias{knit_as_emar}
\title{Compile a RMarkdown into an Explorable Multiverse Analysis Report}
\usage{
knit_as_emar()
}
\description{
Include the function in your setup code chunk to compile a RMarkdown into an Explorable Multiverse Analysis Report
}
\details{
Explorable Multiverse Analysis Reports are interactive documents which allow the user
to interactively explor and look at each distinct end-to-end analysis path within a multiverse analysis.
The idea was put forth by Dragicevic et al. (2019) in \doi{10.1145/3290605.3300295}{\emph{Increasing the 
Transparency of  Research Papers with Explorable Multiverse Analyses}}.
Tangle.js is used to provide interactivity and allow the toggle through different specifications.
}
