% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eng_multidy.R
\name{multidy_engine}
\alias{multidy_engine}
\title{Create custom code chunks for multidy}
\description{
An easier way to interact with a multiverse object by using a custom code engine
}
\details{
This is a custom code engine that allows the users to interface directly with a created multiverse object, 
without using functions such as [inside]. See examples for more details
}
\examples{
\dontrun{
# Typically R users, using RMarkdown could specify code by creating a code chunk, 
# and use the functions provided to add code to the multiverse :
```{r}
M = multiverse()
inside(M, { df = data.frame( x = 1:10 ) })
```

# Here, they would need to reference the multiverse object everytime they want to add anythign to it
# Instead, they could add code to the multiverse by using a custom code engine:
```{multidy, multiverse = M}
df = data.frame( x = 1:10 )
}

}
