% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse.R
\name{multiverse}
\alias{multiverse}
\title{Create a new multiverse object}
\usage{
multiverse()
}
\value{
An empty multiverse object
}
\description{
multiverse() constructs a new multiverse object. The multiverse object is an S3 object wtih three attributes:
\itemize{
  \item code: The `code` attribute stores the user's code for conducting a multiverse analysis. The user can
  define multiple analysis pathways at each step in the analysis using the `branch` call. It supports tidyverse syntax.
  \item parameters: The `parameters` attribute contain the list which itself comprises of two lists: 
  the list of parameters and the list of conditions. The list of parameters is a named list which defines all the values 
  that each defined parameter can take. The list of conditions defines, if any of the parameter values are conditional
   on a specific value of another parameter, the condition. This is created from the `code` attribute.
  \item current_parameter_assignment: The `current_parameter_assignment` attribute is a list which contains a single value
  assignment for each parameter defined in the `code`.
}
}
\examples{
\dontrun{
M <- new_multiverse()
}

}
