% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiverse_code_chunk_addin.R
\name{multiverse_code_chunk_addin}
\alias{multiverse_code_chunk_addin}
\title{Create a multiverse code chunk}
\usage{
multiverse_code_chunk_addin()
}
\description{
Addin for RStudio for creating multiverse code chunks
}
