% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_multiverse.R
\name{parse_multiverse}
\alias{parse_multiverse}
\title{Parse the multiverse syntax to identify branches}
\usage{
parse_multiverse(multiverse, .super_env)
}
\arguments{
\item{multiverse}{The multiverse object with some code passed to it}

\item{.super_env}{The parent environment of the multiverse object i.e. the environment
in which the multiverse object was called. This is automatically recorded.}
}
\value{
The `parse_multiverse` function returns a list of lists. the list of parameters and the list of conditions.
The list of parameters is a named list which defines all the values that each defined parameter can take.
The list of conditions defines, if any of the parameter values are conditional on a specific value of another
parameter, the condition.
}
\description{
In a multiverse, the user can define different values that a parameter can take using the `branch` call.
The `parse_multiverse` identifies the `branch` calls defined in the analysis syntax and parses them into a list of
parameters and the corresponding values that each parameter can take.
}
