% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.multiverse}
\title{Accessing contents of the multiverse object}
\usage{
\method{print}{multiverse}(x, ...)
}
\arguments{
\item{x}{An object of class multiverse}

\item{...}{further arguments passed to or from other methods. Currently ignored.}
}
\description{
prints objects of class multiverse. Provides a quick overview of the multiverse by 
listing the parameters and conditions declared, to get an overview of the number of combinations.
The function outputs upto 20 of the parameters declared along with their option names (upto 5), 
and upto 10 of the conditions declared.
}
