% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{accessors}
\alias{accessors}
\alias{$.multiverse}
\alias{multiverse_table}
\alias{multiverse_table.multiverse}
\alias{code}
\alias{code.multiverse}
\alias{parameters}
\alias{parameters.multiverse}
\alias{conditions}
\title{Accessing contents of the multiverse object}
\usage{
\method{$}{multiverse}(multiverse, name)

multiverse_table(multiverse)

\method{multiverse_table}{multiverse}(multiverse)

code(multiverse)

\method{code}{multiverse}(multiverse)

parameters(multiverse)

\method{parameters}{multiverse}(multiverse)

conditions(multiverse)

conditions(multiverse)
}
\arguments{
\item{multiverse}{Object of class multiverse}

\item{name}{a variable name}
}
\description{
A multiverse object contains several \strong{Object variables}. These can be accessed using convenient functions.
Variables from the analysis that is being performed within the multiverse can be accessed using the `$`.
Object variables such as the `code`, the `multiverse table`, the `parameters` and the `conditions` can be accessed using respective functions
}
