% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{accessors}
\alias{accessors}
\alias{$.multiverse}
\alias{$<-.multiverse}
\alias{expand}
\alias{expand.default}
\alias{expand.multiverse}
\alias{size}
\alias{size.default}
\alias{size.multiverse}
\alias{code}
\alias{code.default}
\alias{code.multiverse}
\alias{parameters}
\alias{parameters.default}
\alias{parameters.multiverse}
\alias{conditions}
\alias{conditions.default}
\alias{conditions.multiverse}
\alias{extract_variable_from_universe}
\title{Accessing contents of the multiverse object}
\usage{
\method{$}{multiverse}(multiverse, name)

\method{$}{multiverse}(multiverse, name) <- value

expand(multiverse)

\method{expand}{default}(multiverse)

\method{expand}{multiverse}(multiverse)

size(multiverse)

\method{size}{default}(multiverse)

\method{size}{multiverse}(multiverse)

code(multiverse, .pretty = TRUE)

\method{code}{default}(multiverse, .pretty = TRUE)

\method{code}{multiverse}(multiverse, .pretty = TRUE)

parameters(multiverse)

\method{parameters}{default}(multiverse)

\method{parameters}{multiverse}(multiverse)

conditions(multiverse)

\method{conditions}{default}(multiverse)

\method{conditions}{multiverse}(multiverse)

extract_variable_from_universe(multiverse, idx, name)
}
\arguments{
\item{multiverse}{Object of class multiverse}

\item{name}{a variable name}

\item{value}{a new value to be assigned}

\item{.pretty}{A binary argument whether `code()` should prettify the output using the tidyverse style guide. defaults to TRUE.}

\item{idx}{index of the universe in the multiverse (corresponds to the row in the table)}
}
\description{
A multiverse object contains several \strong{Object variables}. These can be accessed using convenient functions.
Variables from the analysis that is being performed within the multiverse can be accessed using the \code{$}.
Object variables such as the \code{code}, the \code{expanded parameter options table}, the \code{parameters} and the \code{conditions} can be accessed using respective functions
}
