% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute_default}
\alias{execute_default}
\title{Execute parts of, or the entire multiverse}
\usage{
execute_default(multiverse, N = NA)
}
\arguments{
\item{multiverse}{The multiverse object}

\item{N}{Override the default analysis and instead perform the N-th analysis from the multiverse table}

\item{.vec}{A vector specifying the range of analysis paths from the multiverse to be executed. Defaults to \code{\link[base]{NA}}
which indicates the complete multiverse is to be executed.}
}
\description{
These are functions which allow the user to execute parts or whole of the multiverse. 
The user can choose to either execute the default analysis using the [execute_default], or a part or 
whole of the multiverse using the [execute_multiverse].
}
\details{
Each single analysis within the multiverse lives in a separate environment. We provide convenient functions to execute 
the results for the  default analysis, as well as parts or whole of the multiverse. The default analysis is executed everytime 
code is added to the multiverse, hence we do not expect users to use the [execute_default] function frequently.
}
\examples{
\dontrun{
#' M <- multiverse()
inside(M, {
  data <- rnorm(100, 50, 20)
  
  x.mean <- mean(data, trim = branch(
    trim_values, 
    "trim_none" ~ 0,
    "trim_1pc" ~ 0.05,
    "trim_5pc" ~ 0.025,
    "trim_10pc" ~ 0.05
  ))
})

M \%>\%
  execute_multiverse()
}

}
