% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_json.R
\name{export_2_json}
\alias{export_2_json}
\alias{export_dist_2_json}
\title{Exporting results from a multiverse analysis to JSON}
\usage{
export_2_json(x, term, mean, sd, dist, filename)

export_dist_2_json(x, term, dist, filename)
}
\arguments{
\item{x}{a tidy tibble or data frame which contains summary statistics or distributional information 
of each regression coefficient parameter}

\item{term}{column in the data frame, x, which contains the name of the coefficients}

\item{mean}{column in the data frame, x, which contains the mean estimate for each coefficient}

\item{sd}{column in the data frame, x, which contains the standard error estimate for each coefficient}

\item{dist}{column in the data frame, x, which contains vectorised distributions---an object of class 
`distribution` for each coefficient}

\item{filename}{filename on disk (as a character string)}
}
\description{
export to json
}
