% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside.R
\name{inside}
\alias{inside}
\title{Pass code into the multiverse}
\usage{
inside(multiverse, .expr)
}
\arguments{
\item{multiverse}{A multiverse object. A multiverse object is an S3 object which can be defined using `new_multiverse`}

\item{.expr}{R syntax. All the operations that the user wants to perform within the multiverse can be passed. 
Since it accepts a single argument, chunks of code can be passed using `{}`. See example for details.}
}
\description{
The inside function allows you to define objects and code to manipulate the objects inside the multiverse.
However, the function only stores the code, and does not execute any code at this step. See `eval_multi` for executing the code.
}
